requirements.txt
python-telegram-bot==20.8
python-dotenv==1.0.1
openai==1.11.0
anthropic==0.18.0
aiohttp==3.9.3
asyncio==3.4.3
requests==2.31.0 